feat(spring,gradle): split extension
