+++ concordile-spring/gradle.properties
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-springCloudContractVersion=5.0.0
+springCloudContractVersion=5.0.2
